+++ components/common/index.ts
@@ -1,5 +1,6 @@
+import Button from "./Button";
 import ButtonLink from "./ButtonLink";
 import NavLink from "./NavLink";
 import Container from "./Container";
 
-export { ButtonLink, NavLink, Container }
+export { Button, ButtonLink, NavLink, Container }