+++ components/index.ts
@@ -1,5 +1,5 @@
-import { ButtonLink, NavLink, Container } from './common';
+import { Button, ButtonLink, NavLink, Container } from './common';
 import { Gallery, GalleryImage } from './gallery';
 import { MainNav } from './navigation';
 
-export { ButtonLink, NavLink, Container, Gallery, GalleryImage, MainNav }
+export { Button, ButtonLink, NavLink, Container, Gallery, GalleryImage, MainNav }