+++ lib/mongodb.ts
@@ -1,0 +1,23 @@
+import { MongoClient } from "mongodb";
+
+const uri = process.env.MONGODB_URI!;
+let client: MongoClient;
+let clientPromise: Promise<MongoClient>;
+
+if (!process.env.MONGODB_URI) {
+    throw new Error("Please define the MONGODB_URI environment variable");
+}
+
+if (process.env.NODE_ENV === "development") {
+    // avoid creating multiple clients in dev hot reloads
+    if (!(global as any)._mongoClientPromise) {
+        client = new MongoClient(uri);
+        (global as any)._mongoClientPromise = client.connect();
+    }
+    clientPromise = (global as any)._mongoClientPromise;
+} else {
+    client = new MongoClient(uri);
+    clientPromise = client.connect();
+}
+
+export default clientPromise;